Set `udevadm settle` timeout
On a system, while installing Ubuntu 18.04(.4), all the `udevadm settle` invocations timed out.

It's not clear why this happens, so we set a large enough timeout. On systems without this issue, the timeout won't matter, while on systems with the issue, the timeout will be enough to ensure that the devices are created.