Add support for eMMC disks
Cheap laptops typically have eMMC. Although it's questionable whether ZFS will run on them, it's still worth adding.

This may accidentally give the option to install on external eMMC cards on system having such ports, however, the audience is advanced, so it's ok to display it.

This should also allow to install on SBCs with an eMMC.