Improve encryption passphrase UX (and simplify logic)
The UX has been simplified by implying that the encryption is disabled if the user inputs an empty password (so that there's no need for a screen asking for enablement).

The code follows the same logic, relying on the `ZFS_PASSPHRASE` variable being set or not. This is a different behavior from the other variables, for which blank and unset is the same.