New function invocation mechanism (README!)
The new mechanism is clearly explained in the comment. The rationale is the next implementation of hot swapping.

WATCH OUT! The semantics have changed. Previously, it was evident by looking at the invocation when a function had distro-based overrides; now, that needs to be looked into. Since this behavior is not very impactful, it's not been ported, for simplicity purposes.